@@ -1,6 +1,6 @@
 package de.jthedroid.whatsappchatanalyzer;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.text.DateFormat;
 import java.text.ParseException;
