@@ -1,8 +1,9 @@
 package de.jthedroid.whatsappchatanalyzer;
 
-import android.support.annotation.NonNull;
 import android.util.Log;
 import android.util.LongSparseArray;
+
+import androidx.annotation.NonNull;
 
 import java.io.BufferedReader;
 import java.io.IOException;
