@@ -4,15 +4,16 @@ package de.jthedroid.whatsappchatanalyzer;
 import android.app.DatePickerDialog;
 import android.app.Dialog;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.DialogFragment;
 import android.widget.DatePicker;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
 
 import java.util.Calendar;
 
 
-public class DatePickerDialogFragment extends DialogFragment implements DatePickerDialog.OnDateSetListener {
+class DatePickerDialogFragment extends DialogFragment implements DatePickerDialog.OnDateSetListener {
     private DateReceiver dateReceiver = null;
 
     @NonNull
